% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aligned_map}
\alias{aligned_map}
\title{Scanned and aligned map}
\format{
.jpeg file
}
\source{
'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971)
}
\usage{
aligned_map
}
\description{
Scanned and aligned map from the 'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971).
}
\keyword{datasets}
