% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{border_remove}
\alias{border_remove}
\title{Polygon feature to mask map border}
\format{
Polygon spatial feature, gcs WGS 84 -- WGS84 - World Geodetic System 1984, EPSG:4326
}
\usage{
border_remove
}
\description{
Polygon feature used in the automated cleaning process to mask map border and legend.
}
\keyword{datasets}
