% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_map}
\alias{database_map}
\title{Georeferenced and digitised map}
\format{
Polygon spatial feature, gcs WGS 84 -- WGS84 - World Geodetic System 1984, EPSG:4326
}
\source{
'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971)
}
\usage{
database_map
}
\description{
Georeferenced and digitised map from the 'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971).
}
\keyword{datasets}
