% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixFileNames.R
\name{fixFileNames}
\alias{fixFileNames}
\title{Cange filenames to match Dyntaxa}
\usage{
fixFileNames(path, pattern = "*.geojson")
}
\arguments{
\item{path}{file path folder with range map database}

\item{pattern}{file extension. Default is patterm = "*.geojson"}
}
\value{
new filenames
}
\description{
In the Swedish Taxonomic Database Dyntaxa, hybrid species are denoted by the symbol ×. To facilitate file handling the symbol × has been replaced by the letter x in the database file names. This function changes the letter x in the file names back the symbol × so that the file names are identical to the Dyntaxa scientific names.
}
