% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galium_list}
\alias{galium_list}
\title{List of maps for species in the \emph{Galium} genera}
\format{
List of 14 spatial polygon features, gcs WGS 84 -- WGS84 - World Geodetic System 1984, EPSG:4326
}
\source{
'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971)
}
\usage{
galium_list
}
\description{
List of georeferenced and digitised maps for species in the \emph{Galium} genera from the 'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971).
}
\keyword{datasets}
