% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histPresSweNationalGrid}
\alias{histPresSweNationalGrid}
\title{Historical per grid cell presence of vascular plants in Sweden}
\format{
Data frame, 2,046,758 rows and 3 columns
\describe{
\item{scientificName}{Scientific name of species}
\item{gridID}{Grid ID }
\item{histPres}{1=historical presence}
}
}
\usage{
histPresSweNationalGrid
}
\description{
Historical per grid cell presence in the Swedish 10×10 km National grid of of vascular plants from the 'Atlas of the distribution of vascular plants in northwestern Europe' (Hultén 1971).
}
\keyword{datasets}
