% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_cells}
\alias{land_cells}
\title{IDs of grids with land surface}
\format{
\subsection{\code{land_cells}}{

Character string
}
}
\usage{
land_cells
}
\description{
String of Grid cells IDs with 50\% or more land surface outside Sweden’s national borders and grid cells covered to 80\% or more by sea surface or waterways.
}
\keyword{datasets}
