% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lycopodium_complanatum_00004}
\alias{Lycopodium_complanatum_00004}
\title{Lycopodium complanatum range map}
\format{
Spatial polygon features, gcs WGS 84 -- WGS84 - World Geodetic System 1984, EPSG:4326
}
\usage{
Lycopodium_complanatum_00004
}
\description{
Polygon features of the range of Lycopodium complanatum
}
\keyword{datasets}
