% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCleanConvert.R
\name{mapCleanConvert}
\alias{mapCleanConvert}
\title{Clean raster and convert to polygon feature}
\usage{
mapCleanConvert(x)
}
\arguments{
\item{x}{digitised map (RasterLayer)}
}
\value{
a polygon feature (SpatVector)
}
\description{
Cleans the digitised range map by removing (most of the) map border, removing small inconsistencies by 'clumping' and 'sifting' and converts the raster map to a polygon feature.
}
