% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDig.R
\name{mapDig}
\alias{mapDig}
\title{Digitise scanned range maps}
\usage{
mapDig(x, value = 150)
}
\arguments{
\item{x}{scanned range map (RGB rasterStack).}

\item{value}{maximum pixel value that delimits the pixel values representing the range information in the scanned maps. Default value is x=150.}
}
\value{
A raster layer.
}
\description{
Takes a scanned range map (raster stack) with three bands (Red-Green-Blue) and extracts the range information by taking the green raster band and setting all pixel values up to 150 to 1 and all other to 0. The pixel range 0-150 in green raster band represents the dark red areas in the scanned map.
}
