% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NorthernEurope}
\alias{NorthernEurope}
\title{Map of northern Europe}
\format{
Spatial polygon features, projected gcs SWEREF99 TM, EPSG:3006
}
\usage{
NorthernEurope
}
\description{
Polygon features of countries in Northern Europe
}
\keyword{datasets}
