% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToGrid.R
\name{rangeToGrid}
\alias{rangeToGrid}
\title{Match species' ranges to (national) grid}
\usage{
rangeToGrid(x, pattern = "*.geojson", grid = SweNationalGrid)
}
\arguments{
\item{x}{file path to folder with range map database}

\item{pattern}{file extension. Default is patterm = "*.geojson"}

\item{grid}{(national) grid polygon feature}
}
\value{
a list of data frames with per species per grid cell historical presence
}
\description{
Match species' ranges to (national) grid
}
