% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ScanGrid}
\alias{ScanGrid}
\title{European 10×10 km reference grid over Scandinavia and Finland}
\format{
Spatial polygon feature, projected gcs ETRS89-extended / LAEA Europe, EPSG:3035
}
\source{
https://www.eea.europa.eu/data-and-maps/data/eea-reference-grids-2/gis-files/europe-10-km-100-km
}
\usage{
ScanGrid
}
\description{
Subset of the European 10×10 km reference grid (Scandinavia and Finland).
}
\keyword{datasets}
