% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SweNationalGrid}
\alias{SweNationalGrid}
\title{Swedish 10×10 km national grid}
\format{
Spatial polygon feature, gcs WGS 84 -- WGS84 - World Geodetic System 1984, EPSG:4326
}
\usage{
SweNationalGrid
}
\description{
Swedish 10×10 km national grid
}
\keyword{datasets}
